+++ .patch/prebuild-only-rename.sh
@@ -48,5 +48,10 @@ sed -i \
     -e "s/\"OsmAnd Nightly\"/\"$appName\"/g" \
     "$osmand_dir/build.gradle"
 
+# also rename package
+sed -i \
+    -e "s/\"net.osmand.dev\"/\"net.osmand.djbpf\"/g" \
+    "$osmand_dir/build.gradle"
+
 # return from whence we came (just in case)
 popd